Type func_800428A8's resource argument as the pointer every caller passes
The ninth parameter was declared s32. It is a resource pointer: the function
stores it straight into the object's 0x54 word, which DisplayObject already
declares void *, and func_80042874 immediately above does the same job through
DisplayObjectResource::resource with the parameter spelled void * already.

The compiler had been reporting this from both directions at once, which is
what makes the reading certain rather than plausible. Four resident call sites
passed a pointer and warned "makes integer from pointer without a cast", while
eight others passed the same kind of value through an explicit (s32) or (int)
cast written to satisfy the declaration -- (s32)D_801AF000, (int)D_801AF000 and
(s32)tb, where tb is a u8 * holding D_801AF000 or D_801AF800. Every caller in
the tree, resident and overlay, passes a pointer; only the prototype disagreed.

Correcting the type alone would have moved the problem rather than fixed it:
measured, it cleared four warnings and raised eight new ones in the opposite
direction, because the defensive casts were then casting a pointer to an
integer to pass it as a pointer. The casts are removed with the declaration, so
the whole set goes together: 64 warnings before, 60 after, none of them this
function, and no new warning anywhere.

Gates: resident 84a54ed74f3d0edd6d81380839f7e4ef5bfb21ecea18be9a062bd6bfa5a45c88,
five overlay hashes unchanged, verify-overlays basic-types external-attempts
check-metadata check-candidate-headlines OK.

Related to #2500, #2495

+++ src/game/func_80020F4C.c
@@ -159,7 +159,7 @@ void func_80020F4C(void)
                     obj = func_800400AC(func_8004002C(), 2);
                     func_800428A8(obj, spec->x, spec->y, 0,
                                   gDuel_bWinnerSide, spec->kind, 0x11, 9,
-                                  (s32)D_801AF000);
+                                  D_801AF000);
                     if (obj->field_69 >= 0x1A) {
                         obj->field_40.h.field_40 =
                             (u16)obj->field_40.h.field_40 + 0x10;
